chore: move cookie_options file
