+++ src/undici/cookie_agent.ts
@@ -1,8 +1,9 @@
 import { Agent, errors, Pool } from 'undici';
 
+import type { CookieOptions } from '../cookie_options';
+import { validateCookieOptions } from '../utils/validate_cookie_options';
+
 import { CookieClient } from './cookie_client';
-import type { CookieOptions } from './cookie_options';
-import { validateCookieOptions } from './utils/validate_cookie_options';
 
 type CookieAgentOptions = Agent.Options & {
   cookies?: CookieOptions | undefined;
