+++ src/utils/validate_cookie_options.ts
@@ -1,5 +1,4 @@
 import { CookieJar } from 'tough-cookie';
-import { errors } from 'undici';
 
 import type { CookieOptions } from '../cookie_options';
 
@@ -11,20 +10,18 @@ function validateCookieOptions(
   resolver: ModuleResolver = require.resolve,
 ): asserts opts is CookieOptions {
   if (!(opts.jar instanceof CookieJar)) {
-    throw new errors.InvalidArgumentError('invalid cookies.jar object');
+    throw new TypeError('invalid cookies.jar object');
   }
 
   if (opts.async_UNSTABLE) {
     try {
       resolver('deasync');
     } catch (_err) {
-      throw new errors.InformationalError('you should install deasync library when cookies.async_UNSTABLE is true.');
+      throw new Error('you should install deasync library when cookies.async_UNSTABLE is true.');
     }
   } else {
     if (!opts.jar.store.synchronous) {
-      throw new errors.InvalidArgumentError(
-        'you should set cookies.async_UNSTABLE to true for using the asynchronous cookie store.',
-      );
+      throw new TypeError('you should set cookies.async_UNSTABLE to true for using the asynchronous cookie store.');
     }
   }
 }
